Improving the performance of online applications.
1. The cache mechanism of the contracts master was abolished, and the
information of contracts were managed in contract's blocks to reduce the
conflict between threads.

2. In previous versions, it was slow that the process of converting a
phone number from a long value to a string. As a countermeasure for the
slow conversion process, the conversion result was cached.

As a result of the improving performance, the size of this cache grew
and adversely affected the performance.

This cache mechanism was removed and the conversion process was modified
to a logic that was fast enough even without the cache mechanism.

3. Improved  performance of the CreateTestData command.We have applied
the same contract information management mechanism to the CreateTestData
command as to the online applications.