Add configuration options for load testing in Altimeter
burstDurationSec: This parameter defines the duration of burst mode. Its
default value is set to 0, indicating that burst mode is disabled by
default.

normal.duration.sec: This specifies the duration (in seconds) for the
normal mode. The default value for this parameter is set to 30 seconds.

txs.per.session: This parameter sets the number of transactions (TX) to
be executed in a single session. In online applications, the session is
refreshed after executing the number of transactions specified by this
parameter. This functionality is active only when using Iceaxe. The
default value is 0, which means the session will not refresh.